Move various bloom gateway metrics into separate file
This commit introduces a new metrics struct that contains both the
metrics specific the server/handler and the metrics specific to the
worker.

This has the advantage that metrics can be instantiated and accessed
from one single place.
