+++ pkg/bloomgateway/worker.go
@@ -8,8 +8,6 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/services"
 	"github.com/pkg/errors"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/common/model"
 
 	"github.com/grafana/loki/pkg/queue"
@@ -24,65 +22,6 @@ const (
 
 type workerConfig struct {
 	maxItems int
-}
-
-type workerMetrics struct {
-	dequeueDuration   *prometheus.HistogramVec
-	processDuration   *prometheus.HistogramVec
-	metasFetched      *prometheus.HistogramVec
-	blocksFetched     *prometheus.HistogramVec
-	tasksDequeued     *prometheus.CounterVec
-	tasksProcessed    *prometheus.CounterVec
-	blockQueryLatency *prometheus.HistogramVec
-}
-
-func newWorkerMetrics(registerer prometheus.Registerer, namespace, subsystem string) *workerMetrics {
-	labels := []string{"worker"}
-	r := promauto.With(registerer)
-	return &workerMetrics{
-		dequeueDuration: r.NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "dequeue_duration_seconds",
-			Help:      "Time spent dequeuing tasks from queue in seconds",
-		}, labels),
-		processDuration: r.NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "process_duration_seconds",
-			Help:      "Time spent processing tasks in seconds",
-		}, append(labels, "status")),
-		metasFetched: r.NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "metas_fetched",
-			Help:      "Amount of metas fetched",
-		}, labels),
-		blocksFetched: r.NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "blocks_fetched",
-			Help:      "Amount of blocks fetched",
-		}, labels),
-		tasksDequeued: r.NewCounterVec(prometheus.CounterOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "tasks_dequeued_total",
-			Help:      "Total amount of tasks that the worker dequeued from the queue",
-		}, append(labels, "status")),
-		tasksProcessed: r.NewCounterVec(prometheus.CounterOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "tasks_processed_total",
-			Help:      "Total amount of tasks that the worker processed",
-		}, append(labels, "status")),
-		blockQueryLatency: r.NewHistogramVec(prometheus.HistogramOpts{
-			Namespace: namespace,
-			Subsystem: subsystem,
-			Name:      "block_query_latency",
-			Help:      "Time spent running searches against a bloom block",
-		}, append(labels, "status")),
-	}
 }
 
 // worker is a datastructure that consumes tasks from the request queue,
